lel shells

+++ install.sh
@@ -1,27 +1,27 @@
 
 #!/bin/sh
 install_vim_config() {
-    rm -rf ~/.config/nvim
+    [ -d "~/.config/nvim" ] && rm -rf ~/.config/nvim
     cp -r $DOTFILES/nvim/ ~/.config/nvim
 }
 
 install_tmux_config() {
-    rm ~/.tmux.conf
+    [ -f "~/.tmux.conf" ] && rm ~/.tmux.conf
     cp $DOTFILES/tmux/.tmux.conf ~/.tmux.conf
 }
 
 install_alacritty_config() {
-    rm -rf ~/.config/alacritty
+    [ -d "~/.config/alacritty" ] && rm -rf ~/.config/alacritty
     cp -r $DOTFILES/alacritty ~/.config/alacritty
 }
 
 install_zsh_config() {
-    rm ~/.zshrc
+    [ -f "~/.zshrc" ] && rm ~/.zshrc
     cp $DOTFILES/zsh/.zshrc ~/.zshrc
 }
 
 install_fish_config() {
-    rm -rf ~/.config/fish
+    [ -d "~/.config/fish" ] && rm -rf ~/.config/fish
     cp -r $DOTFILES/fish/ ~/.config/fish
 }
